requirements.txt
### dependency
tqdm
dvc
tensorflow
pandas
numpy
Scipy
pyyaml
scikit-learn
lxml
matplotlib
### local packages -
-e . # << uncomment this to install from setup.py
